1 step forward, 10 steps back.
Embedding Request/Response as simple enum types isn't quite enough. The
command's paramter should be bounded to the type itself to ensure that
request/responses are properly formed.

To this end, the enum type should capture the parameter values, rather
than having a top level Message struct with both an Command enum and a
Params vector.